+++ NEWS.md
@@ -10,9 +10,10 @@ This is a leader-intensive update. Additions/new features include:
 - `gml_part`: a data set on participant summary information from the GML MID data.
 - `add_gml_mids()` now works with leader-year data. See documentation for more.
 - `add_spells()` will do more general peace spell calculations, though this package will now depend on v. 1.3.0 of `{stevemisc}`.
-- `add_lwuf()` adds estimates of leader willingness to use force to leader-year analyses.
+- `add_lwuf()` adds estimates of leader willingness to use force to leader-year and leader-dyad-year analyses.
 - `add_lead()` adds some leader attributes to leader-year analyses.
 - `gml_mid_dirleaderdisps` and `gml_mid_ddlydisps` include data on leader dyads in conflict.
+- `declare_attributes()` allows users to declare package-specific attributes to outside data they bring to the R session. This allows the user to use many of the shortcuts in this package to their data.
 
 Other changes include:
 
@@ -25,6 +26,7 @@ its original size. This is a huge reduction in disk space. The original simulati
 - `add_peace_years()` has been superseded by `add_spells()`. The "superseded" language communicates the function will still work as it has for previous releases, for data types supported by the function, though its development will stop. `add_spells()` promises to have greater functionality.
 - `archigos` data's `ccode` has been renamed to `gwcode` to reflect these are actually Gleditsch-Ward state codes.
 - The "whittle" functions now work with leader-dyadic conflict data available in the package.
+
 
 peacesciencer 0.6.0
 ---------------------------------------------------------------------